Silence numeric conversion warnings in eig and tensor code
Tighten numeric casts across the eigensolver, MPO, tensor, and iterative solver code paths to silence conversion warnings without changing the underlying algorithms.

This also improves float128 and random-number utilities by constraining conversions more explicitly, adding complex random generation helpers, and using deterministic random initialization where Eigen's built-in Random() was a poor fit for the supported scalar types.

+++ source/math/eig/solver_spectra/solver_spectra.impl.h
@@ -99,10 +99,10 @@ void eig::solver_spectra<MatrixType>::eigs() {
         result.meta.rows           = matrix.rows();
         result.meta.cols           = eigvecs_spectra.cols();
         result.meta.nev            = nev_internal;
-        result.meta.nev_converged  = eigvals.size(); // Holds the number of converged eigenpairs on exit
+        result.meta.nev_converged  = static_cast<int>(eigvals.size()); // Holds the number of converged eigenpairs on exit
         result.meta.ncv            = ncv_internal;
         result.meta.tol            = static_cast<double>(tol);
-        result.meta.iter           = sol.num_iterations() + sol.num_operations();
+        result.meta.iter           = safe_cast<int>(sol.num_iterations() + sol.num_operations());
         result.meta.num_mv         = matrix.num_mv;
         result.meta.num_pc         = matrix.num_pc;
         result.meta.num_op         = matrix.num_op;
